Reworked index for page transitions
