+++ src/transitions.scss
@@ -1,17 +1,21 @@
 .page-enter {
-  opacity: 0.01;
+  opacity: 0;
+  transform: scale(1.1);
 }
 
 .page-enter-active {
   opacity: 1;
-  transition: opacity 500ms ease-in;
+  transform: scale(1);
+  transition: opacity 300ms, transform 300ms;
 }
 
 .page-exit {
   opacity: 1;
+  transform: scale(1);
 }
 
 .page-exit-active {
-  opacity: 0.01;
-  transition: opacity 300ms ease-in;
+  opacity: 0;
+  transform: scale(0.9);
+  transition: opacity 300ms, transform 300ms;
 }
